Adds type of trigger to event name
Slight refactor
